removes the vo_person_id variable names (becomes username)
In the claims returned by CILogon, the former OIDC Provider,
'vo_person_id' was key whose value became the username of
the django.contrib.auth.User instance created in both HERMES
and SCiMMA Auth. In the move to SCiMMA's own OIDC Provider
(the Keycloak instance at login.scimma.org), that key became
'sub'. But it's a username and this refactor creates consistent
argument and variable names reflecting that.