populate topics filter choices at runtime
populating the choices dynamically needs to be moved to the
__init__() so it happens at runtime, not "compile"-time. (The
specific failure was running the 'migrate' management command
in the 'apply-db-migrations' init container during deployment.
But basicly, the project couldn't be interpreted b/c the table
for the Message model object (hermes_message) hasn't been created
but the interpreter was trying to access it).